fix: getContent fetching only 1 entity
Today I find out that XF does a quite smart thing - it fetches content for Alerts in bulk, for each content type. That's something I was not ready for, and probably they changed somewhere in 2.2\2.3, because it was working fine for me in 2.1 (it was working, right?).

Anyway, now It's fixed and I should take a note and probably use this pattern in my other projects...